requirements.txt
typer
phi
phi2
phidata
streamlit
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
sentence-transformers
InstructorEmbedding
ollama
google-generativeai